Standardize data collection task screen and field naming conventions
- Rename various `*TaskFragment.kt` files to `*TaskScreen.kt` for Consistency (Number, MultipleChoice, CaptureLocation, Text, DrawArea, DropPin, Instruction).
- Refactor `PhotoTaskFragment.kt` and `PhotoTaskScreen.kt` into a unified `PhotoTaskScreen.kt` and move the UI content to `PhotoTaskContent.kt`.
- Refactor `TimeTaskFragment.kt` and `TimeTaskScreen.kt` into a unified `TimeTaskScreen.kt` and move the input field to `TimeTaskField.kt`.
- Rename `DateTaskScreen` to `DateTaskField` and update `DateTaskFragment` to use the new component name.
- Update `PhotoTaskScreenTest` to reflect the renaming of `PhotoTaskScreen` to `PhotoTaskContent`.
