+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/location/CaptureLocationTaskScreen.kt
