+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/point/DropPinTaskScreen.kt
