+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/polygon/DrawAreaTaskScreen.kt
