+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/time/TimeTaskScreen.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2026 Google LLC
+ * Copyright 2023 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,73 +15,79 @@
  */
 package org.groundplatform.android.ui.datacollection.tasks.time
 
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.interaction.PressInteraction
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
+import android.app.TimePickerDialog
+import android.content.Context
+import android.content.DialogInterface
+import android.text.format.DateFormat
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import org.groundplatform.android.ui.common.ExcludeFromJacocoGeneratedReport
-import org.groundplatform.ui.theme.AppTheme
-
-const val TIME_TEXT_TEST_TAG: String = "time task input test tag"
-
-// TODO: Add trailing icon (close logo) for clearing selected time.
+import androidx.compose.ui.platform.LocalContext
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import java.text.SimpleDateFormat
+import java.util.Calendar
+import java.util.Date
+import org.groundplatform.android.R
+import org.groundplatform.android.model.submission.DateTimeTaskData
+import org.groundplatform.android.ui.datacollection.DataCollectionViewModel
+import org.groundplatform.android.ui.datacollection.tasks.TaskContainer
+import org.groundplatform.ui.theme.sizes
 
 @Composable
-fun TimeTaskScreen(
-  timeText: String,
-  hintText: String,
-  onTimeClick: () -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  val interactionSource = remember { MutableInteractionSource() }
-  LaunchedEffect(interactionSource) {
-    interactionSource.interactions.collect { interaction ->
-      if (interaction is PressInteraction.Release) {
-        onTimeClick()
-      }
+fun TimeTaskScreen(viewModel: TimeTaskViewModel, dataCollectionViewModel: DataCollectionViewModel) {
+  val taskData by viewModel.taskTaskData.collectAsStateWithLifecycle()
+  val context = LocalContext.current
+
+  val timeText =
+    remember(taskData) {
+      (taskData as? DateTimeTaskData)?.let {
+        DateFormat.getTimeFormat(context).format(Date(it.timeInMillis))
+      } ?: ""
+    }
+
+  val hintText = remember {
+    val timeFormat = DateFormat.getTimeFormat(context)
+    if (timeFormat is SimpleDateFormat) {
+      timeFormat.toPattern().uppercase()
+    } else {
+      "HH:MM AM/PM" // Fallback hint if DateFormat is not SimpleDateFormat
     }
   }
 
-  Column(modifier = modifier) {
-    // TODO: Replace with simple text field.
-    OutlinedTextField(
-      value = timeText,
-      onValueChange = {},
-      readOnly = true,
-      placeholder = { Text(hintText) },
-      modifier = Modifier.width(200.dp).testTag(TIME_TEXT_TEST_TAG),
-      interactionSource = interactionSource,
+  TaskContainer(viewModel = viewModel, dataCollectionViewModel = dataCollectionViewModel) {
+    TimeTaskField(
+      modifier = Modifier.padding(horizontal = MaterialTheme.sizes.taskViewPadding),
+      timeText = timeText,
+      hintText = hintText,
+      onTimeClick = { showTimeDialog(context, viewModel) },
     )
   }
 }
 
-@Preview(showBackground = true)
-@Composable
-@ExcludeFromJacocoGeneratedReport
-private fun TimeTaskScreenPreview() {
-  AppTheme {
-    Column(
-      modifier = Modifier.padding(16.dp),
-      verticalArrangement = Arrangement.spacedBy(8.dp),
-      horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-      TimeTaskScreen(timeText = "", hintText = "HH:MM AM", onTimeClick = {})
-      Spacer(modifier = Modifier.height(10.dp))
-      TimeTaskScreen(timeText = "10:30 AM", hintText = "HH:MM AM", onTimeClick = {})
-    }
+fun showTimeDialog(context: Context, viewModel: TimeTaskViewModel) {
+  val calendar = Calendar.getInstance()
+  val hour = calendar[Calendar.HOUR]
+  val minute = calendar[Calendar.MINUTE]
+  val timePickerDialog =
+    TimePickerDialog(
+      context,
+      { _, updatedHourOfDay, updatedMinute ->
+        val c = Calendar.getInstance()
+        c[Calendar.HOUR_OF_DAY] = updatedHourOfDay
+        c[Calendar.MINUTE] = updatedMinute
+        viewModel.updateResponse(c.time)
+      },
+      hour,
+      minute,
+      DateFormat.is24HourFormat(context),
+    )
+  timePickerDialog.setButton(DialogInterface.BUTTON_NEUTRAL, context.getString(R.string.clear)) {
+    _,
+    _ ->
+    viewModel.clearResponse()
   }
+  timePickerDialog.show()
 }
